Fix controller failing when implementing billing methods other than Sisow.
When adding payment methods than the Sisow::Billing ones, controller failed.
The controller called a message on the billing_method that most don't
have. And that should not be called even if they have it.
This is in added filter
Hence we guard against this by checking for Sisow billing methods only.